+++ tests/Integration/DSC_SqlLogin.Integration.Tests.ps1
@@ -514,30 +514,78 @@ try
                 } | Should -Not -Throw
             }
 
-            It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -Be 'True'
+            It 'Should have set the resource and all the parameters should match' {
+                $resourceCurrentState = $script:currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName `
+                    -and $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.Ensure | Should -Be 'Present'
+                $resourceCurrentState.Name | Should -Be $ConfigurationData.AllNodes.DscUser4Name
+                $resourceCurrentState.LoginType | Should -Be $ConfigurationData.AllNodes.DscUser4Type
+                $resourceCurrentState.LoginPasswordExpirationEnabled | Should -BeTrue
+                $resourceCurrentState.LoginPasswordPolicyEnforced | Should -BeFalse
             }
 
-            It 'Should allow SQL Server, login username and password to connect to SQL Instance (using SqlConnection.Open())' {
-                $serverName = $ConfigurationData.AllNodes.ServerName
-                $instanceName = $ConfigurationData.AllNodes.InstanceName
-                $databaseName = $ConfigurationData.AllNodes.DefaultDbName
-                $userName = $ConfigurationData.AllNodes.DscUser4Name
-                $password = $ConfigurationData.AllNodes.DscUser4Pass1 # Original password
-
-                $sqlConnectionString = 'Data Source={0}\{1};User ID={2};Password={3};Connect Timeout=5;Database={4};' -f $serverName, $instanceName, $userName, $password, $databaseName
-
-                {
-                    $sqlConnection = New-Object System.Data.SqlClient.SqlConnection $sqlConnectionString
-                    $sqlConnection.Open()
-                    $sqlConnection.Close()
-                } | Should -Not -Throw
+            It 'Should return $true when Test-DscConfiguration is run' {
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
         Wait-ForIdleLcm -Clear
 
-        $configurationName = "$($script:dscResourceName)_UpdateLoginDscUser4_UpdateLoginDscUser4_Config_LoginPasswordPolicyEnforced"
+        # Reset LoginPasswordExpirationEnabled and LoginPasswordExpirationEnabled flags
+        $configurationName = "$($script:dscResourceName)_UpdateLoginDscUser4_Config"
+
+        Context ('When using configuration {0}' -f $configurationName) {
+            It 'Should compile and apply the MOF without throwing' {
+                {
+                    $configurationParameters = @{
+                        OutputPath                 = $TestDrive
+                        # The variable $ConfigurationData was dot-sourced above.
+                        ConfigurationData          = $ConfigurationData
+                    }
+
+                    & $configurationName @configurationParameters
+
+                    $startDscConfigurationParameters = @{
+                        Path         = $TestDrive
+                        ComputerName = 'localhost'
+                        Wait         = $true
+                        Verbose      = $true
+                        Force        = $true
+                        ErrorAction  = 'Stop'
+                    }
+
+                    Start-DscConfiguration @startDscConfigurationParameters
+                } | Should -Not -Throw
+            }
+
+            It 'Should be able to call Get-DscConfiguration without throwing' {
+                {
+                    $script:currentConfiguration = Get-DscConfiguration -Verbose -ErrorAction Stop
+                } | Should -Not -Throw
+            }
+
+            It 'Should have set the resource and all the parameters should match' {
+                $resourceCurrentState = $script:currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName `
+                    -and $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.Ensure | Should -Be 'Present'
+                $resourceCurrentState.Name | Should -Be $ConfigurationData.AllNodes.DscUser4Name
+                $resourceCurrentState.LoginType | Should -Be $ConfigurationData.AllNodes.DscUser4Type
+                $resourceCurrentState.Disabled | Should -Be $false
+                $resourceCurrentState.LoginMustChangePassword | Should -Be $false # Left the same as this cannot be updated
+                $resourceCurrentState.LoginPasswordExpirationEnabled | Should -Be $false
+                $resourceCurrentState.LoginPasswordPolicyEnforced | Should -Be $false
+            }
+        }
+
+        Wait-ForIdleLcm -Clear
+
+        $configurationName = "$($script:dscResourceName)_UpdateLoginDscUser4_Config_LoginPasswordPolicyEnforced"
 
         Context ('When using configuration {0} (to update back to original password)' -f $configurationName) {
             It 'Should re-compile and re-apply the MOF without throwing' {
@@ -569,24 +617,21 @@ try
                 } | Should -Not -Throw
             }
 
-            It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -Be 'True'
+            It 'Should have set the resource and all the parameters should match' {
+                $resourceCurrentState = $script:currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName `
+                    -and $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.Ensure | Should -Be 'Present'
+                $resourceCurrentState.Name | Should -Be $ConfigurationData.AllNodes.DscUser4Name
+                $resourceCurrentState.LoginType | Should -Be $ConfigurationData.AllNodes.DscUser4Type
+                $resourceCurrentState.LoginPasswordExpirationEnabled | Should -BeFalse
+                $resourceCurrentState.LoginPasswordPolicyEnforced | Should -BeTrue
             }
 
-            It 'Should allow SQL Server, login username and password to connect to SQL Instance (using SqlConnection.Open())' {
-                $serverName = $ConfigurationData.AllNodes.ServerName
-                $instanceName = $ConfigurationData.AllNodes.InstanceName
-                $databaseName = $ConfigurationData.AllNodes.DefaultDbName
-                $userName = $ConfigurationData.AllNodes.DscUser4Name
-                $password = $ConfigurationData.AllNodes.DscUser4Pass1 # Original password
-
-                $sqlConnectionString = 'Data Source={0}\{1};User ID={2};Password={3};Connect Timeout=5;Database={4};' -f $serverName, $instanceName, $userName, $password, $databaseName
-
-                {
-                    $sqlConnection = New-Object System.Data.SqlClient.SqlConnection $sqlConnectionString
-                    $sqlConnection.Open()
-                    $sqlConnection.Close()
-                } | Should -Not -Throw
+            It 'Should return $true when Test-DscConfiguration is run' {
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }